+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - full screen view menu option added to project window
 - full screen shortcut is Ctrl + Cmd + F or F11 on Windows (#57)
 - more shortcuts added for zoom in/out: +, =, Ctrl-+ and Ctrl-= could all zoom in, -, Ctrl-- could all zoom out (#57)
+- added option to automatically run retouch after run. The option can be persisted in the settings (#56)
 
 ----